Created first edition of GUI-Installer
*Massive* overhaul from the command-line installer
Instead of having to manually type in each program, this allows for the user to just select the programs they want and click "Enter" to Download/Install everything
Added some initial comments to make the code easier to follow.
When clicking the All/None/Defaults buttons, the checkboxes will actually update accordingly
Current issues:
It was originally my goal to have the command line within the GUI to allow for users to still input the commands they were used to. This may still be put into place, but it's currently commented out for the time being.
Similar story for having a progress bar to track the downloads. This may still be added, but it's also commented out for the time being.
The "Delete Setup Files" button is non-functional. This should fixed in a future commit.
Trying to download Adobe Reader will often cause the program to hang. It seems like it usually still works in the long run, but I'd like to improve this.
There are still artifacts of the old command-line code. I'm unsure how to best handle this. Some of the functions may become buttons and some of the code may just be deleted.
There's currently a blank space between the top and bottom chunks of the GUI. I'm unsure what to put there - to be determined at another time.
The GUI is nice and relatively clean, but I think the aesthetics of it could definitely be improved.
I'm hoping that this will ultimately replace the original script. I may keep the old one for archiving purposes, but I'm unsure if it'll be updated further unless there's something I feel is a necessary change.